+++ .replit
@@ -1,6 +1,5 @@
 language="python3"
 run = """
-pip install -U git+https://github.com/Pycord-Development/pycord
 python main.py
 """
